% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sf2arc.R
\name{sf2arc}
\alias{sf2arc}
\title{Converts an \code{sf} object to an ArcGIS feature class}
\usage{
sf2arc(sf_object, fc_path)
}
\arguments{
\item{fc_path}{character; Path to the ArcGIS feature class.}

\item{sp_object}{\code{sf} object}
}
\value{
Writes the \code{sf} object to an ArcGIS feature class specified
    by path.
}
\description{
Converts an \code{sf} object to an ArcGIS feature class in
    the specified geodatabase feature class.
}
\details{
The \code{sf2arc} function requires the prior installation of the
\code{arcgisbinding} package AND a licensed installation of ESRI
\code{ArcGIS Desktop} or \code{ArcGIS Pro}. The \code{arcgisbinding}
package can be installed from within \code{ArcGIS Pro} or by following the
instructions at
\url{https://r-arcgis.github.io/assets/arcgisbinding-vignette.html} for
\code{ArcGIS Desktop} users.
}
\references{
\describe{
  \item{ESRI ArcGIS Desktop, ArcGIS Pro}{\url{https://pro.arcgis.com/}}
  \item{\code{arcgisbinding}}{
  \url{https://r-arcgis.github.io/assets/arcgisbinding-vignette.html}}
}
}
\seealso{
The \code{\link{arc2sf}} function for loading data from and ESRI
spatial dataset.
}
